Fix memory leak on error path of connect
We kept around the result of getaddrinfo and assigned the first
sockaddr_in we could connect to the conn. If the device isn't able to
reach PH but is alive after a while we would leak the getaddrinfo result
returned previously since we never free the old result.

make a copy of the sockaddr we require instead of keeping the one from
getaddrinfo. If we're not able to make the copy we just lookup one from
getaddrinfo again and try to make a copy of that one.