Add central resolver capability for PV
This patch makes the PH connection as an attribute of the PV and passes
this information about the socket down to the thttp when attempting to
connect with PH.

Also fix the caching of already looked up ph api URL. Earlier we were
always flushing the cache and freeing up the cached sockaddr.

TODO:
Should thttp do a lookup on its own if connectivity drop after PV has
been initialized.