Make lxc-logs appear in separate log files
This patch
1. introduces lxc log per container started in a separate log file.

2. makes default log directory as
/storage/logs/<container_name>/<container_name>.log

3. Honors lxc.conf setting over the default above.

Since lxc_log_init and do_lxcapi_load_config both try to initialize the
log, and use a static global variable to check if the log has been
initialized or not, we need to fork first before setting up the lxc-log
and starting the container.

The started container pid is returned via pipe to the parent.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>