make tsh_run wait properly and return exit status
This patch makes tsh_run wait for the command to exit by first disabling
the SIGCHLD and then doing a fork. Also adds one extra parameter to
return back the exit code of the command it waited for to complete.

Also added a log entry while trying to zip the previous log file. Though
recursive function call but it should be ok unless log size is way too
small.

pthread_cond_signal is not a signal safe function and there was also a
problem of priority inversion in guaranteeing that signal handler sees
the pid to wait on so removed that code.