Always post update status on first boot
This patch sends the status of device as UPDATED if it boots properly.
It could be possible that during update we miss or do a power cycle and
never able to send out the update status as updated but the device would
boot with the correct revision. If the device booted, then send out the
update status as UPDATE_DONE only once with the revision value stored in
current state of the device.