Update device status properly after update
This patch updates the device status post an update by keeping a track
of whether or not the status was posted and if not then it remembers the
current status of device post update and keeps trying to update the
status until it succeeds or another reboot occurs.

There was also wrong argument order while trying to set status in
pv_set_current_status. This patch fixes that too.

Signed-off-by: Pranay Kumar Srivastava <pranay.srivastava@pantacor.com>