Fix pv_log_flush
This patch fixes pv_log_flush by correctly taking into account the last
JSON "]", if it can be placed or otherwise sacrifice 1 byte of log to
complete the json.

The problem occurs when there are too many log items with level DEBUG
and no level < DEBUG in between. When that occurs the ring buffer fills
up with too many entries. These entries are then tried to be pushed to
PH at one go whenever an entry whole level is < DEBUG arrives, which
happens when trying to do an update, which causes multiple buffers to be
allocated which are then filled with the log values.

The older version didn't correctly check for the last , closing, brace
']' if it can be added or not to the current json buffer which caused
pantavisor to crash whenever there were too many entries whose
collective buffer required multiple buffers to be allocated.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>