Add delay to open_socket when retrying to connect
The function used to sit in a tight infinite loop until the connect
system call returned success. Now the function has been changed to
try to connect 5 times with a delay of 2 seconds. If this is not
successful the function returns -errno to signify that the file
descriptor returned from this function is not usable.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>