Fix log buffer check
This patch changes how the ring buffer is checked when adding a new
entry in it. The check of whether or not the buffer has been filled is
now done first before trying to add a new entry. Not doing this caused
the pointer to move past the allocated buffer and caused crash.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>