Remove newlines at the end of pv_log
This patch removes the extraneous line endings which might be added by
using pv_log in pantavisor. The patch doesn't remove the line-endings in
a generic way from log_add since the log can be from platform and we
could potentially remove the newline added by the platform in it's log.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>