Change logger attribute to trusted
Some file systems viz tmpfs do not support user attributes. Since we run
as a priviliged containers it's ok to use "truseted.*" namespace. This
is done to avoid having xattr errors when trying to save the current
logger position as an xattr value with the file attributes.

NOTE: When using overlayfs, the upper dir's file system must support the
trusted xattr namespace and the container must be running as a
priviliged container.

TODO:
For unprivliged containers the "trusted" xattr namespace may not work.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>