Add revision retry capability in pantavisor.
If a revision fails to download then pantavisor just reports an error
that it wasn't able to download. The download issue could be due to a
network connectivity and therefore the device should retry the download
of the revision instead of reporting an error on the first try.

Introduce two new parameters in pantavisor config. These parameters need
to be present in pantavisor.config

* revision.retries which is the number of retries to do in case of a
  download failure.
* revision.retries.timeout which is the timeout value in seconds before
  actually attempting to retry downloading that revision.

Signed-off-by: Pranay Kumar Srivastava <pranay.srivastava@pantacor.com>