Update user meta when keys are removed
This patch properly removes the user meta keys and the associated files
when user meta key is removed.

Earlier the code used to clear the meta keys before going for the new
keys. This meant that we lost the ability to check which keys were found
for user-meta and which ones were deleted.

The changes done,
1. First Update the in-memory structure with a flag PV_USERMETA_ADD if
   the meta key was found.

2. Remove the keys which were not found and also remove the associated
file.

Signed-off-by: Pranay Kumar Srivastava <pranay.srivastava@pantacor.com>