Commit update after a specified duration
A new config parameter updater.commit.delay is added through which one
can control the delay a device needs to be "up" for the new revision to
be committed and accepted as the current revision for the device.
The duration must be supplied in seconds and the default duration is 300
seconds.
This is a ph config item,
updater.commit.delay = <a positive value>

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>