Add download progress message to PH
This patch introduces object download progress information upload to PH.
The objects which needs to be downloaded are only shown as part of the
progress field as shown below,

"progress": 0,
  "downloads": {
    "total": {
      "object_name": "total",
      "object_id": "none",
      "total_size": 314572800,
      "start_time": 1598440086,
      "current_time": 1598440094,
      "total_downloaded": 0
    },
    "objects": [
      {
        "object_name": "dumdum/dummy9",
        "object_id": "e3b0c44298",
        "total_size": 0,
        "start_time": 1598440093,
        "current_time": 1598440094,
        "total_downloaded": 0
      }
    ]

Objects array may or may not appear depending if the device was able to
allocate enough space for it, however the total object would always be
present.

The start time and current time are both relative to the device.
The progress for an object being downloaded is uploaded every 3 seconds
however it's always "attempted (as mentioned above)" to be uploaded when
an object is correctly downloaded.
This upload delay can't be configured currently.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>