upload meta information from platform to PH
Each platform will have a directory pv-plat-meta where any files created
/ modified or deleted would be used to upload the meta data as contents
of that file. The key would be the plat-name/<file-name> with contents
fully json escaped.

When the device comes up the device-meta data is checked from PH and
files of format <plat_name>/<filename> are checked for being present on
the plat-meta directory. If the filename isn't found the contents are
cleared from the device-meta for this file.

The plat-meta is mounted as a rw bind mount. On host the plat-meta
directory is present as /pv-plat-meta. Since /pv is mounted as ro we use
another directory here and since we've ramfs as rootfs the files are not
persistent.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>