perf: introduce event loop
Main loop is now an event loop, using libevent. All states are triggered
by a callback as soon as transitioned into, except PV_STATE_WAIT, which
is called periodically every 2 seconds.

The function that reads, processes and replies to pv-ctrl HTTP requests
is now also a callback. This was made to remove the 2 second select and
leave libevent some room to loop and execute other events.

Finally, device metadata PUT and user metadata GET requests to Pantacor
Hub are also triggered as events, as well as the precessing of their
replies.