fix: not all _config files visible in container namespace
Duplicated files were wrongly committed when installing a new revision
to the trail directory. This caused some config files that have the
same content to not show up in the container mount namespace.

The fix keeps a record of all object install locations and uses that
list to install the revision directory.